A program that orgnizies all your loose MP3 files into folders with the name of the artist, album and year.
This is a program written in C that uses a unix enviornment to take an
optional single argument, path to a directory contating a collection of
MP3 files, and operates on it. The program scans the music collection
and arranges the MP3 files in a collection by artist, album and year and
moves the files into the corresponding folders.